[ADD] hooks - useViewModeWithSider for main page refactoring
